Minor changes (#204)
* chore(config): remove unused threshold property

* chore(api/search): use `nn_count`

* chore: truncate len 500 -> 100 when seaching

* chore(test-script): specify `nn_count`

* interm changes

+++ src/models/rpc.rs
@@ -27,7 +27,7 @@ pub struct VectorANN {
     pub vector_db_name: String,
     pub vector: Vec<f32>,
     pub filter: Option<Filter>,
-    pub nn_count: Option<i32>,
+    pub nn_count: Option<usize>,
 }
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
@@ -35,7 +35,7 @@ pub struct BatchVectorANN {
     pub vector_db_name: String,
     pub vectors: Vec<Vec<f32>>,
     pub filter: Option<Filter>,
-    pub nn_count: Option<i32>,
+    pub nn_count: Option<usize>,
 }
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
